布置作业删除已布置作业

+++ src/main/java/com/niitcoder/coursegrade/service/impl/CourseHomeworkServiceImpl.java
@@ -1,5 +1,10 @@
 package com.niitcoder.coursegrade.service.impl;
 
+import com.niitcoder.coursegrade.domain.CourseAttachment;
+import com.niitcoder.coursegrade.domain.StudentHomework;
+import com.niitcoder.coursegrade.repository.CourseAttachmentRepository;
+import com.niitcoder.coursegrade.repository.StudentHomeworkRepository;
+import com.niitcoder.coursegrade.security.SecurityUtils;
 import com.niitcoder.coursegrade.service.CourseHomeworkService;
 import com.niitcoder.coursegrade.domain.CourseHomework;
 import com.niitcoder.coursegrade.repository.CourseHomeworkRepository;
@@ -11,7 +16,6 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -25,9 +29,13 @@ public class CourseHomeworkServiceImpl implements CourseHomeworkService {
     private final Logger log = LoggerFactory.getLogger(CourseHomeworkServiceImpl.class);
 
     private final CourseHomeworkRepository courseHomeworkRepository;
+    private final StudentHomeworkRepository studentHomeworkRepository;
+    private final CourseAttachmentRepository courseAttachmentRepository;
 
-    public CourseHomeworkServiceImpl(CourseHomeworkRepository courseHomeworkRepository) {
+    public CourseHomeworkServiceImpl(CourseHomeworkRepository courseHomeworkRepository, StudentHomeworkRepository studentHomeworkRepository, CourseAttachmentRepository courseAttachmentRepository) {
         this.courseHomeworkRepository = courseHomeworkRepository;
+        this.studentHomeworkRepository = studentHomeworkRepository;
+        this.courseAttachmentRepository = courseAttachmentRepository;
     }
 
     /**
@@ -75,20 +83,48 @@ public class CourseHomeworkServiceImpl implements CourseHomeworkService {
      * @param id the id of the entity.
      */
     @Override
-    public void delete(Long id) {
+    public void delete(Long id) throws Exception {
         log.debug("Request to delete CourseHomework : {}", id);
-        courseHomeworkRepository.deleteById(id);
+
+        //当前id对应的作业是否是当前登录用户的
+        Optional<CourseHomework> courseHomework = courseHomeworkRepository.findById(id);
+        if (courseHomework.isPresent()) { //id是否可以查询到作业
+            CourseHomework homework = courseHomework.get();
+            String loginName = SecurityUtils.getCurrentUserLogin().get();
+            if (!homework.getPlan().getCourse().getCourseUser().equals(loginName)) {
+                throw new Exception("无权限删除此课程.");
+            }
+            //检查是否有学生提交作业
+            List<StudentHomework> studentHomeworks = studentHomeworkRepository.findByHomeworkId(id);
+            if (studentHomeworks != null && !studentHomeworks.isEmpty()) {
+                throw new Exception("作业已经有提交记录");
+            }
+            //检查作业是否有附件
+            List<CourseAttachment> courseAttachments = courseAttachmentRepository.findByHomeworkId(id);
+            if (courseAttachments != null && !courseAttachments.isEmpty()) {
+                throw new Exception("作业已有附件提交");
+            }
+            courseHomeworkRepository.deleteById(id);
+
+        } else { //id对应的课程不存在
+            throw new Exception("作业不存在");
+        }
+
+
+        /**
+         * Get one courseHomework by id.
+         *
+         * @param id the id of the entity.
+         * @return the entity.
+         */
+
+
     }
-    /**
-     * Get one courseHomework by id.
-     *
-     * @param id the id of the entity.
-     * @return the entity.
-     */
+
+
+
     @Override
     public List<CourseHomework> findByPlanId(Long id) {
-        List<CourseHomework> courseHomework = new ArrayList<>();
-        courseHomework = courseHomeworkRepository.findByPlanId(id);
-        return courseHomework;
+        return null;
     }
 }
